+++ app/src/main/java/com/kickstarter/mock/services/MockApolloClient.kt
@@ -20,7 +20,7 @@ import com.kickstarter.models.Comment
 import com.kickstarter.models.CreatorDetails
 import com.kickstarter.models.ErroredBacking
 import com.kickstarter.models.Location
-import com.kickstarter.models.PostCommentInput
+import com.kickstarter.services.mutations.PostCommentData
 import com.kickstarter.models.Project
 import com.kickstarter.models.Reward
 import com.kickstarter.models.StoredCard
@@ -73,7 +73,7 @@ open class MockApolloClient : ApolloClientType {
         )
     }
 
-    override fun createComment(comment: PostCommentInput): Observable<Comment> {
+    override fun createComment(comment: PostCommentData): Observable<Comment> {
         return Observable.just(CommentFactory.comment())
     }
